pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "llm-agent"
version = "0.1.0"
description = "A CLI LLM agent with context strategies and multi-provider support (Qwen / OpenAI / Claude)"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "tiktoken>=0.7",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
llm-agent = "llm_agent.interfaces.cli.main:main"
chat = "llm_agent.interfaces.cli.interactive_strategies:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["llm_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
